+++ src/action/index.ts
@@ -1,0 +1,3 @@
+import { run } from './action';
+
+export const action = { run };
