--- src/actions-lib.ts
@@ -1,7 +1,0 @@
-import * as core from '@actions/core';
-import * as github from '@actions/github';
-
-type Core = typeof core;
-type Github = typeof github;
-
-export { core, github, Core, Github };
