+++ src/lib.ts
@@ -1,0 +1,7 @@
+import * as core from '@actions/core';
+import * as github from '@actions/github';
+
+export type Core = typeof core;
+export type Github = typeof github;
+
+export { core, github };
